requirements.txt
asn1crypto==0.24.0
azure-common==1.1.23
azure-storage-blob==2.0.1
azure-storage-common==2.0.0
boto3==1.9.197
botocore==1.12.197
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
docutils==0.14
future==0.17.1
idna==2.8
ijson==2.4
jmespath==0.9.4
pycparser==2.19
pycryptodomex==3.8.2
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
snowflake-connector-python==1.8.5
urllib3==1.25.3
